+++ SERVER.CLASS.BUILDER/Thread_Locker_Builder/Thread_Locker_Builder.cpp
@@ -159,7 +159,7 @@ void Thread_Locker_Builder::Build_Thread_Locker(){
 
 void Thread_Locker_Builder::Determine_Compiler_Command(){
 
-     char Process_Command [] = {'g','+','+',' ','-','c',' ','-','s','t','d','=','c','+','+','1','4',' ','\0'};
+     char Process_Command [] = {'g','+','+',' ','-','c',' ','-','s','t','d','=','c','+','+','1','7',' ','\0'};
 
      char Source_File_Name [] = "Thread_Locker.cpp";
 
@@ -169,13 +169,13 @@ void Thread_Locker_Builder::Determine_Compiler_Command(){
 
      char Include_Word [] = "-include";
 
-     char Thread_Library_Name [] = {'-','l','p','t','h','r','e','a','d','\0'};
+     char Thread_Library_Name [] = "-lwinpthread";
 
      char Space_Character [] = {' ','\0'};
 
      char Output_Redirection_Command [] = {'2','>',' ','\0'};
 
-     char Error_Message_File_Name [] = {'/','C','o','m','p','i','l','e','r','_','O','u','t','p','u','t','\0'};
+     char Error_Message_File_Name [] = {'\\','C','o','m','p','i','l','e','r','_','O','u','t','p','u','t','\0'};
 
      int Source_File_Name_Size = strlen(Source_File_Name);
 
@@ -236,11 +236,22 @@ void Thread_Locker_Builder::Determine_Compiler_Command(){
      this->Compiler_Command[index_counter] = '\0';
 }
 
+void Thread_Locker_Builder::Build_Output_Stream_File(){
+
+     std::string path = "Compiler_Output";
+
+     this->FileManager.SetFilePath(path);
+
+     this->FileManager.FileOpen(RWCf);
+
+     this->FileManager.FileClose();
+}
+
 void Thread_Locker_Builder::Remove_Header_Extra_File(){
 
      char Header_Extra_File [] = "Thread_Locker.h.gch";
 
-     char Directory_Character [] = {'/','\0'};
+     char Directory_Character [] = {'\\','\0'};
 
      char * Current_Directory = this->Directory_Manager.GetCurrentlyWorkingDirectory();
 
@@ -262,14 +273,29 @@ void Thread_Locker_Builder::Remove_Header_Extra_File(){
 
      File_Path[index_counter] = '\0';
 
-     this->FileManager.Delete_File(File_Path);
+     if(this->FileManager.Is_Path_Exist(File_Path)){
+
+        this->FileManager.Delete_File(File_Path);
+     }
 
      delete [] File_Path;
 }
 
 void Thread_Locker_Builder::Run_System_Commands(){
 
-     int system_return_value =  this->System_Interface.System_Function(this->Compiler_Command);
+     this->Build_Output_Stream_File();
+
+     /*
+
+     std::cout << "\n Inside Thread_Locker_Builder::Run_System_Commands";
+
+     std::cout << "\n this->Compiler_Command:" << this->Compiler_Command;
+
+     std::cin.get();
+
+     */
+
+     int system_return_value =  system(this->Compiler_Command);
 
      if(system_return_value != 0 ){
